--- src/assets/styles/pages/MainPage/searchBoxView.scss
@@ -1,20 +1,0 @@
-.searchBox {
-  height: 48px;
-  display: flex;
-  gap: 16px;
-  text-align: center;
-  margin: 0 192px 64px 192px;
-
-  .searchBox-input {
-    width: 90%;
-    padding: 16px 0 16px 64px;
-    border-radius: 8px;
-  }
-
-  .searchBox-submit {
-    width: 10%;
-    border-radius: 8px;
-    background-color: #404040;
-    color: white;
-  }
-}
